Login Page Base Implementation
I have recreated the basic design of the login  page and have had to make some changes to the design because I couldn't get it to work without center alignment. Added nav tag for group of links as per project requirement. Plus, I added comments for both home page and login page as I realized code looked to monotone and was difficult to digest and change afterwards.